refs #53 Updated the loading and parsing unit tests.
